+++ libs/traits/Cargo.toml
@@ -13,6 +13,7 @@ documentation.workspace = true
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+chrono = { workspace = true }
 cfg-primitives = { workspace = true }
 frame-support = { workspace = true }
 impl-trait-for-tuples = { workspace = true }
