+++ libs/types/src/pools.rs
@@ -12,7 +12,7 @@
 
 use cfg_traits::{
 	fee::{FeeAmountProration, PoolFeeBucket},
-	Seconds,
+	time::Seconds,
 };
 use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
 use scale_info::TypeInfo;
