+++ libs/types/src/portfolio.rs
@@ -10,7 +10,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-use cfg_traits::{Seconds, TimeAsSecs};
+use cfg_traits::time::{Seconds, TimeAsSecs};
 use frame_support::{pallet_prelude::RuntimeDebug, traits::Get, BoundedVec};
 use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
 use scale_info::TypeInfo;
